On ia64 time_t is 64 bit. Explicitly cast tv_sec to long and change the corresponding format specifier to %ld in a call to printf() in function softclock(). The printf() is conditional upon DIAGNOSTIC.
Found by: LINT


git-svn-id: svn+ssh://svn.freebsd.org/base/head@119358 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/kern/kern_timeout.c
@@ -231,9 +231,9 @@ softclock(void *dummy)
 					maxdt = bt2.frac;
 					bintime2timespec(&bt2, &ts2);
 					printf(
-			"Expensive timeout(9) function: %p(%p) %d.%09ld s\n",
+			"Expensive timeout(9) function: %p(%p) %ld.%09ld s\n",
 					c_func, c_arg,
-					ts2.tv_sec, ts2.tv_nsec);
+					(long)ts2.tv_sec, ts2.tv_nsec);
 				}
 #endif
 				if (!(c_flags & CALLOUT_MPSAFE))
